genai: add File API (#85)
Add support for the file API, which allows uploading large files
to the service so they can be used to generate content.

The GetFile, DeleteFile and ListFiles RPCs are straightforward.
We use the same GAPIC client for them as we have been using for the
rest of the API.

However, the CreateFile RPC involves a media upload, which GAPICs
do not support. We rely instead on the discovery client, which
we have to generate ourselves. That involves a shell script
to do the work, and a copy of an internal package (gensupport)
from the repo that contains the discovery clients.